Add Payment/ View Payment
The base for both Add Payment and View Payment is in place. The Add Payment is graphically (the view is up and running) functional but the back end and connecting to the API is not working. The base for View Payment, with stubbed data, is in place but for some reason it cannot locally locate files that exist. Extended the error message out to the team and waiting for reply.